+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/controller/SyncTaskController.java
@@ -23,8 +23,8 @@ import info.avalon566.shardingscaling.core.config.SyncType;
 import info.avalon566.shardingscaling.core.synctask.RealtimeDataSyncTask;
 import info.avalon566.shardingscaling.core.execute.Event;
 import info.avalon566.shardingscaling.core.execute.EventType;
-import info.avalon566.shardingscaling.core.execute.executor.SyncJobExecutor;
-import info.avalon566.shardingscaling.core.execute.executor.local.LocalSyncJobExecutor;
+import info.avalon566.shardingscaling.core.execute.engine.SyncJobExecutor;
+import info.avalon566.shardingscaling.core.execute.engine.local.LocalSyncJobExecutor;
 import info.avalon566.shardingscaling.core.execute.reader.LogPosition;
 import info.avalon566.shardingscaling.core.execute.reader.ReaderFactory;
 import info.avalon566.shardingscaling.core.execute.util.DataSourceFactory;
